3: [Susana] Moving blueprint definition to be recognized by routes, fix 404
