+++ book-domain-service/app/routes.py
@@ -2,6 +2,9 @@ from flask import request, Response
 from flask import current_app
 from app.book import Book
 from app import db
+from flask import Blueprint
+
+checkout_blueprint = Blueprint('checkout_page', __name__)
 
 def checkout_book(id):
 
@@ -14,12 +17,12 @@ def checkout_book(id):
         return Response("I wish we had that book! Try another", 404)
 
 
-@current_app.route('/<int:id>/checkout', methods=['PUT'])
+@checkout_blueprint.route('/<int:id>/checkout', methods=['PUT'])
 def checkout(id):
 
     return checkout_book(id)
 
-@current_app.route('/addbook', methods=['POST'])
+@checkout_blueprint.route('/addbook', methods=['POST'])
 def add_book():
     new_title = request.json['title']
