+++ book-domain-service/app_service.py
@@ -1,5 +1,4 @@
 from app import create_app
-
 from config import Config
 
 app = create_app(Config)
